Sweep dead exports and dedupe HelpDocName + META_LINE
