+++ src/shared/path.ts
@@ -8,11 +8,3 @@
 export function toPosix(path: string): string {
   return path.includes('\\') ? path.replace(/\\/g, '/') : path;
 }
-
-/** Last segment of a forward-slash path. Convenience for renderer code
- *  that previously did `p.split('/').pop()` — keeps the same behaviour
- *  but spells the intent. */
-export function posixBasename(path: string): string {
-  const i = path.lastIndexOf('/');
-  return i === -1 ? path : path.slice(i + 1);
-}
